pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "specagent"
version = "0.1.0"
description = "Agentic RAG system for 3GPP telecommunications specifications"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = [
    "rag",
    "llm",
    "3gpp",
    "telecommunications",
    "langgraph",
    "agents",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core LLM & Agent Framework
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.1.0",
    
    # Vector Store & Embeddings
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    
    # API & Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    
    # Data Processing
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "huggingface-hub>=0.26.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
    "typer>=0.13.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.34.0",
    "pytest-mock>=3.14.0",
    
    # Code Quality
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "bandit>=1.7.0",
    "pre-commit>=4.0.0",
    
    # Type Stubs
    "types-requests>=2.32.0",
    "pandas-stubs>=2.2.0",
]

eval = [
    # Evaluation Frameworks
    "ragas>=0.2.0",
    "arize-phoenix>=5.0.0",
    "openinference-instrumentation-langchain>=0.1.0",
    
    # LangChain Evaluation
    "langsmith>=0.1.0",
]

ui = [
    # Demo Interface
    "gradio>=5.6.0",
]

all = [
    "specagent[dev,eval,ui]",
]

[project.scripts]
specagent = "specagent.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/specagent"
Documentation = "https://github.com/yourusername/specagent#readme"
Repository = "https://github.com/yourusername/specagent"
Issues = "https://github.com/yourusername/specagent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/specagent"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

# ============================================================================
# Pytest Configuration
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--cov=src/specagent",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=70",
]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may require mocked services)",
    "e2e: End-to-end tests (full pipeline)",
    "slow: Slow tests (skip with -m 'not slow')",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

# ============================================================================
# Coverage Configuration
# ============================================================================
[tool.coverage.run]
source = ["src/specagent"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

# ============================================================================
# Ruff Configuration (Linting & Formatting)
# ============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004"]
"src/specagent/cli.py" = ["B008"]  # function call in default argument (typer pattern)

[tool.ruff.lint.isort]
known-first-party = ["specagent"]

# ============================================================================
# MyPy Configuration
# ============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "faiss",
    "sentence_transformers",
    "langgraph.*",
    "langchain.*",
    "langchain_community.*",
    "langchain_huggingface.*",
    "ragas.*",
    "phoenix.*",
    "gradio.*",
]
ignore_missing_imports = true

# ============================================================================
# Bandit Configuration (Security)
# ============================================================================
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]  # assert statements (used in tests)
